Remove ability to turn off qthreads work stealing (for good this time)
The ability to turn off work stealing has been added and removed twice now.
Both times to try out the performance of the sherwood scheduler with work
stealing off. The first time it was turned off unconditionally and the second
only if an env var was set.

We are going to use the nemesis scheduler by default and only use sherwood for
numa now. I'm removing this not because we're using nemesis by default as it
might have been useful for numa but because it doesn't do what we expect it to.
It seems sherwood is highly intertwined with work stealing. I believe that when
you create tasks it places them all on a local shepherd and relies on work
stealing to distribute work (where as nemesis or something deals out work to
shepherds in a round robin fashion.) So turning off work stealing the way we
were didn't have the desired behavior.