Add missing generated variables to runtime/include/chplcgfns.h
runtime/include/chplcgfns.h is home to extern variables and function
declarations that are defined in the generated code. It hasn't been updated in
a while so I added all of the current ones and added a note in codegen that new
values should be placed in this file.

I ran into this when I wanted to check CHPL_STACK_CHECKS. I didn't originally
realize this file existed and while adding CHPL_STACK_CHECKS I figured I'd
update all of them.