Clean up in a few spots.
In chplsys, for Linux and Cygwin define chpl_getNumCoresOnThisNode() in
terms of the number of PUs returned by chpl_getNumPUsOnThisNode().

In flat/LocaleModel.chpl initialize numCores with the value returned by
chpl_getNumPUsOnThisNode(), which returns what used to be returned by
chpl_numCoresOnThisLocale().  Thus the value of numCores will be the
same as it was before all this work.