diff --git a/compiler/include/flags_list.h b/compiler/include/flags_list.h index 19a77d3..2a10ad5 100644 --- a/compiler/include/flags_list.h +++ b/compiler/include/flags_list.h @@ -205,6 +205,9 @@ symbolFlag( FLAG_TYPE_CONSTRUCTOR , npr, "type constructor" , ncm )  symbolFlag( FLAG_TYPE_VARIABLE , npr, "type variable" , "contains a type instead of a value" )  symbolFlag( FLAG_USER_NAMED , npr, "user named" , "named by the user" /* so leave it alone */ )  symbolFlag( FLAG_VIRTUAL , npr, "virtual" , ncm ) +// Used to mark where a compiler generated flag was removed (but is desired +// elsewhere). +symbolFlag( FLAG_WAS_COMPILER_GENERATED, npr, "was compiler generated", "used to be marked compiler generated")  symbolFlag( FLAG_WIDE , npr, "wide" , ncm )  symbolFlag( FLAG_WIDE_CLASS , npr, "wide class" , ncm )  symbolFlag( FLAG_WRAPPER , npr, "wrapper" , "wrapper function" ) Pick up remaining test cases in lydia/userDefinedDefaultFunction/
Only 4 of the 9 test cases were being handled correctly, because we key off of whether the
resolved default constructor is compiler-generated, and only replace the default
constructor call with a call to _defaultOf() in that case.
In general, the creation of a wrapper drops the compiler-generated flag, so test cases
involving generic types were failing.  I worked around this by supplying a new flag to
indicate that a default constructor was originally compiler-generated.  We can convert the
default constructor call to a _defaultOf() call if either of the two flags is true.