Fix four test cases failing due to early (double) deletion.
After getting all the userDefinedDefaultFunction test cases working, there were four test
cases failing due to early deallocation errors.  I traced this to the fact that
_defaultOf() acts like a constructor.  So I added it to the set of cases where the return
value should not be auto-destroyed.

I decided against adding a flag to support this, because I expect this particular hack to
be temporary.  When _defaultOf() is a method, it will not behave like a constructor.