Initial commit to try and get rid of qthread deadlock
There are a couple of tests that deadlock with qthreads now that the nemesis
scheduler is the default. They are tests where the main task is in a loop
waiting for a sync var to have a certain value and other tasks are being
created inside the while loop to increment the sync var. Some of the tasks
getting created never get to run because the main task never yields and thus
the main loop is waiting for tasks that it's not letting run to update a value.
To remedy this the plan is to do a qthread_yield every x unlocks for any given
sync var. This does a qthread_yield every 1024 iterations. I tried with 32 and
256 and they both seemed to yield too often through the non-scientific method
of putting some prints when we yield and seeing how often we yield for a few
test programs and also checking how long they take to finish into account.