Revised the output of failing test case distributions/deitz/test_distribution_syntax2.
Since user default constructor calls supplant value-initialization (_defaultOf) in this
patch, it is likely that more field-initialization code is required to support correct
initialization in the user-defined constructor for Block and Cyclic distributions.

It's not really worth fixing this problem right now, as changes in the constructor story
are anticipated which will remedy this problem "for free".