Skip parallel/taskPar/private on cygwin due to hangs.
`test/parallel/taskPar/sungeun/private.chpl` intermittently hangs in nightly
cygwin testing. Normally, this would just result in a timeout, but the
timedexec script is (for some reason unknown to me) unable to kill the
test. Disabling this for now until someone can investigate the hang and/or why
timedexec cannot kill the test when it does hang.