Remove arbitrary limit on command line length
This commit is a minor variation on one contributed by Greg Titus
of Cray.  The purpose is to remove the arbitrary 1024-character
limit on the length of the ssh command line run via system() when
AMUDP is using SPAWNFN=S.  This allows for a much longer command
line, as required by the Chapel runtime when using this spawner.