Revert "Change qthreads topology/affinity setup to work with nemesis"
This reverts commit 27d267d7e6167b11a11a66a5d07a9335f841a2e8.

This commit was intended to fix the number of shepherds and workers per
shepherd that qthreads used for single threaded schedulers like nemesis.
While it corrected the number, it did it after the affinity layer was setup
which meant that we had threads being poorly pinned to cores.

Now the chpl_task_init (or an expert user) is responsible for setting the right
number of shepherds and workers per shepherd before qthreads init, so we no
longer need this.