Add comments to copyPropagation.cpp about failures when scalar propagation is disabled.
Notes that the implementation of reference tracking is incomplete, because references are
not tracked through record fields.

Probably the best way to address this weakness is to eliminate the PRIM_SET_MEMBER and
PRIM_GET_MEMBER_VALUE primitives.  These can be implemented using address extraction
(PRIM_ADDR_OF and PRIM_GET_MEMBER) and PRIM_ASSIGN.