Fix previous attempt to get the lcd architecture for chapel modules
Not only do we need to get the lcd in printchplenv, the other chplenv scripts
that look for the existence of files in the gen directory need to use it as
well.

Note that even in places that don't really need the lcd such as chpl_gmp I
still changed it because it would be a source of confusion if for instance we
defaulted to our regular gmp defaults on cray instead of using system..

I noticed this on an xe where it couldn't find the right gen path because it
detected re2 was none because it was looking in a gen directory for the current
arch instead of the lcd one.