Make qthreads our default tasking layer
I believe we have all the correctness and performance regressions out of the
way and we are ready to switch to using qthreads as the default tasking layer.

This makes that change for all cases except arch == knc or platform == cygwin

This does NOT update our testing scripts. For tonight we will have multiple
machines testing the same configurations. This switch is to see if there are
any surprises or if anything with qthreads blows up. If it does it's likely
we'll just back it out.

If everything goes well, I'll update the testing scripts tomorrow. As a
temporary measure to make sure gasnet testing goes well, this just copies
setting oversubscription and turning off affinity in gasnet common.

Lastly this turns the previous scripts that were testing qthreads performance
alongside fifo to now test qthreads+tcmalloc. Eventually they should be testing
fifo, but we want to get some performance numbers for tcmalloc.