Fix config_param.future: Internal error in getVisibilityScope().
This test failed because a type definition inside a param "if" block is not actually in
the AST.  But an attempt is made to resolve the type constructor anyway in
resolveAutoCopies().

The cure is to weed out this undesirable case.

The .good file had to be updated ("false" is expected).

The .future file was removed, as the test now passes.