Removing Srini from CONTRIBUTORS.devel
[trivial, not reviewed]

Sung's recent mention of CONTRIBUTORS.devel reminded me of its existence
and made me wonder who was in it.  Srini at some point contributed to
code that made it into the compiler and was promoted to CONTRIBUTORS,
but was not removed from this file.  Doing that now.

[As a historical reminder: This file was created to capture people who
contributed code to the repository, but whose contributions were never
included in an official release]