First attempt to fix launchcmd for PBSPro machines
XE testing has been failing saying that aprun doesn't have enough resources.
for PBSPro this is because when you do a qsub, you have to give it the number
of nodes you want and the number of cores per node. This is what we do in the
pbs-aprun launcher. This adds code to the launchcmd script, that mirrors how we
detect the number of cpus to reserve in the pbs launcher.