Quiet TAGS complaints when building Qthreads
We've had some noise in the runtime's build of the TAGS file due
to the fact that the generic glob patterns that were designed to
cover the header files for the task/ and thread/ layers didn't
match any files.

The fix I've taken here is to move the specification of the .h
files out of the tag command itself, in Makefile.help, and into
the Makefile.include files in question.  So far I've only done
it for the default (Qthreads) configuration, which seems to be
working.

Next step is to do the same for other task/thread options, and
arguably for other aspects of the runtime that utilize header
files.