Unify stress and stress.numthr execopts and turn off guard pages
There are two versions of this test because we don't have a way to use a
different execenv for a single test. I believe they are intended to be
identical so I made the execopts the same.

This also turns off guard pages for both as it is a stress test that creates a
large number of tasks and takes a long amount of time to complete with guard
pages. stress.numthr already had guard pages off but as a compopt instead of
as part of the execenv.

The compopts is nice because it's general and the way we want users to do it,
but the execenv is nice because you don't have to copy it for each compopt. For
now I decided to stick with using QT_GUARD_PAGES=false over the compopt, and
once CHPL_STACK_CHECKS also backs whether guard pages are used at execution time
I will replace all uses of QT_GUARD_PAGES