Updated README to reflect new "Quick Start" workflow and refer to PERFORMANCE
The main change here is to swap steps 2 and 3.  The reason we did the
make before the environment previously was so that (a) because we
could (and it validated that the make worked without any specific
settings) and (b) one could say "You only need to do steps 3 onward
once the initial build has been done" (though the insertion of the
make check step arguably broke this as well).

Added two new steps at the end of the Quick Start list, one that tells
how to build the preferred configuration; a second that points to the
PERFORMANCE file in hopes of avoiding the dreaded --fast question.