Rename and change functionality of runtime function for allocating a local string.
I never liked the name initString, but I found a better name for it
due to restructuring the functionality.  It's now called stringMove(),
which copies the len bytes of src into dest and returns dest.  If dest
is "", then it allocates a new buffer for the copy, and returns that
buffer.  It is the caller's responsibility to make sure the src is
large enough to hold dest.

This puts the buffer reuse policy into the new string module.  Note
that the allocation must be done in the runtime since there is no way
to call chpl_here_alloc() from the runtime.