Build fewer LLVM components to speed compile
This patch significantly reduces the build
size and the build time. It omits some LLVM
and Clang components we don't use and only
builds host and x86 and x86-64 backends.

Used to take 23 minutes, now takes 18 minutes
to CHPL_LLVM=llvm time make on my machine. In
addition, llvm/build is now 8G (was 12G)
and llvm/install is now 4.6G (was 5.8G).