Stopped treating Samples/ as special in the test system
The Samples/ directory was, historically, meant to be a set of
illustrative tests that were not tested by default; however,
as a result of not being tested, they also became stale and
disused.  Here, I'm repurposing the Samples/ directory to be
tested to keep it live, so am not having start_test skip over
it by default as it used to.  Updated documentation to reflect
this.