Updating README and testReleaseHelp to include 'rehash' for csh users
Sadly, csh users who set their environment first and build second (as
we now require in the quickstart world) need to do a rehash in order
to update their path to see 'chpl'.  Added a step for them in the README
and updated the testReleaseHelp script to do the same thing.