Fix callStackSize for qthreads
Previously callStackSize only checked for the qthreads environment variables
that control guard pages to determine the size of the call stacks. However, now
that we control them through the --no-stack-checks/--stack-checks options (and
other options that imply them.) It seems like a lot of work to check for the
different options and maintain precedence level and everything. Instead this
just sets QT_INFO=1. If guard pages are on the message "QTHREADS: Guard Pages
Enabled" will print. This handles all cases, including if qthreads was built
without guard pages. It hard codes the message qthreads prints, but this is no
more fragile than coding in the guard page size.