Split the test classes/bradc/arrayInClass/genericArrayInClass-otharrs into a test and a future.
The test shows the ability to create and iterate over arrays-of-arrays that are stored in
classes (which is a valuable test by itself).

The future is due to a bug in zippered iterator inlining, which makes the test fail unless
inlining is turned off.

So, iterator inlining is turned off in the .compopts file associated with the main test.
Details of the bug are provided in the .bad and .future files in the added future test.