Squash line numbers when comparing .bad and .comp.out.tmp files.
This patch is intended to cut down on some testing noise.

If tests report errors relative to module code and the module code has changed for other
reasons, line numbers in the .bad file will get out of sync.  This modification of
sub_test normalizes line numbers out of the comparison, under the assumption that
the exact line numbers in module code are not of particular interest.

This normalization affects only comparison with the .bad file (not the .good file).

Line numbers for errors arising in user (non-module) code are still matched verbatim.

To identify error messages, the sed script matches only lines containing
"CHPL_HOME/modules".  It then replaces the first string of digits between two colons with
"nnnn".