Simple cleanup -- remove original listdir()
The original listdir() was well past its point of usefulness, so
removed it now that findfiles() replaced its functionality and
renamed listdir2() to listdir().

This also meant that the code put into the Filerator help module
(originally because it was in two separate places) was now only
used in one, so removed the helper module (which never worked
anyway) and moved its contents into listdir().