Rewrote fannkuchs to use unbounded ranges as originally hoped for
In investigating this syntax error yesterday, I realized that the
cause was not because the parser does not support unbounded ranges in
an iterator context (whew!) but rather that there's an ambiguity
between the upper bound being a domain literal and the curly bracket
that starts a compound statement (shoot!).  To that end, I found two
ways to work around the issue (parenthesize the unbounded range or use
the 'do' keyword) and updated the future to note it and to request
whether there might be a way to change the precedence table to "do
the right thing" (on the argument that domain literals do not make
good upper bounds for ranges).