Use output file specified by user instead of appending %j.out.
Update slurm-srun and slurm-gasnetrun_ibv launchers to treat
CHPL_LAUNCHER_SLURM_OUTPUT_FILENAME env var as full file path. They both use
the value directly as the --output argument when calling sbatch.

Previously, these launchers would append "%j.out" to the user specified value
and use that for --output argument. This is unnecessary (and slightly
restrictive) because the user can specify a value with escaped characters, like
%j, that slurm will populate with the appropriate values.

Confirmed with @ronawho that this is an ok change.