Remove pointer to STATUS for range iterator overflow
I'm not quite sure where the best place to put the note is, but STATUS isn't
quite right. Something about what the behavior of iteration over ranges is
needed in the spec, but I'm not sure if we want to point to that, or if there
should be a technote or something else.