Solve numa error with memory/shannon/printFinalMemStat
Pointed out in the release status meeting as failing recently, this test has
actually been failing since July 11th, 2014.  The failure mode was a positive
change in the amount of memory allocated, possibly due to Michael's cache
commit on the 10th (though tracking that down has proved really painful).  It
seems correct to just update the .good file with the new numbers.