Remove REGRESSIONS entries for memory/shannon/printFinalMemStat
The regressions file had noted that Michael's cache commit increased memory
usage by 8 bytes on July 11th for no-local and numa, leaving the test failing.
On August 4th, the no-local entry became no longer relevant because Tom's fix
to replicateGlobalWrappedRecordVars() improved memory usage by 64 bytes and
Tom updated the no-local.good file to reflect this.  Numa testing also improved
by this amount, but that was not noticed at the time.  The numa .good file has
now been altered to pass testing, so this entry will not be observed again
except by chance.  Though it is important to note that these changes occurred,
I do not feel it is useful to leave them in the REGRESSIONS file to be cleaned
up when the explanation is no longer known.